refactor : refactor create method in menu
