--- src/main/java/kitchenpos/menus/domain/MenuProduct.java
@@ -1,64 +1,0 @@
-package kitchenpos.menus.domain;
-
-import kitchenpos.products.tobe.domain.entity.Product;
-
-import javax.persistence.*;
-import java.util.UUID;
-
-@Table(name = "menu_product")
-@Entity
-public class MenuProduct {
-    @Column(name = "seq")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    private Long seq;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(
-        name = "product_id",
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_menu_product_to_product")
-    )
-    private Product product;
-
-    @Column(name = "quantity", nullable = false)
-    private long quantity;
-
-    @Transient
-    private UUID productId;
-
-    public MenuProduct() {
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public void setProduct(final Product product) {
-        this.product = product;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(final long quantity) {
-        this.quantity = quantity;
-    }
-
-    public UUID getProductId() {
-        return productId;
-    }
-
-    public void setProductId(final UUID productId) {
-        this.productId = productId;
-    }
-}
