+++ src/main/java/kitchenpos/menus/infra/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
 
+import kitchenpos.menus.tobe.domain.entity.Menu;
+import kitchenpos.menus.tobe.domain.repository.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
